renamed card and cards-list components to place-*
